+++ libs/rxjs/random/src/index.ts
@@ -3,15 +3,25 @@
  *
  * @packageDocumentation
  * @module Random
- *
  */
 /* istanbul ignore file */
+
+// Math.random
 export { fromRandom } from './lib/from-random';
 export { fromRandomInt } from './lib/from-random-integer';
 export { fromRandomStr } from './lib/from-random-string';
+
+// Web Crypto
+export { fromRandomBytes } from './lib/from-random-bytes';
 export { fromRandomCrypto } from './lib/from-random-crypto';
+export { fromRandomCryptoCharset, fromRandomCryptoStr } from './lib/from-random-crypto-str';
+export { fromRandomCryptoInt } from './lib/from-random-crypto-int';
+export { fromRandomUUID } from './lib/from-random-uuid';
 export { fromUUIDv4 } from './lib/from-uuid-v4';
 
-export { FromRandomStringOpts } from './types/from-random-string';
-export { FromRandomCryptoOpts } from './types/from-random-crypto';
+// Types
+export type { FromRandomCryptoOpts } from './types/from-random-crypto';
+export type { FromRandomStringOpts } from './types/from-random-string';
+
+// Defaults
 export { RND_STR_DEFAULTS } from './utils/from-random-string';