+++ libs/rxjs/random/src/utils/from-uuid.ts
@@ -10,10 +10,12 @@
  * @private
  * @internal
  */
+import { getCrypto } from './get-crypto';
+
 export function uuidv4() {
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore
   return ([1e7] + -1e3 + -4e3 + -8e3 + -1e11).replace(/[018]/g, (c) =>
-    (c ^ (crypto.getRandomValues(new Uint8Array(1))[0] & (15 >> (c / 4)))).toString(16),
+    (c ^ (getCrypto().getRandomValues(new Uint8Array(1))[0] & (15 >> (c / 4)))).toString(16),
   );
 }